+++ src/Configuration/SkConfigEntry.cs
@@ -14,6 +14,7 @@ namespace SkToolbox.Configuration
         public static ConfigEntry<bool> CAllowChatOutput { get; set; }
         public static ConfigEntry<bool> CAllowExecuteOnClear { get; set; }
         public static ConfigEntry<bool> CKeepAchievements { get; set; }
+        public static ConfigEntry<bool> CGodModeBlocksDamage { get; set; }
         public static ConfigEntry<bool> CConsoleEnabled { get; set; }
         public static ConfigEntry<bool> CScrollable { get; set; }
         public static ConfigEntry<int> CScrollableLimit { get; set; }